Finished raw parser

+++ src/currex/structure/primitives/FloatPrimitive.java
@@ -2,7 +2,7 @@ package currex.structure.primitives;
 
 import currex.visitor.Visitor;
 
-public record FloatPrimitive(Float value) implements Primitive {
+public record FloatPrimitive(Double value) implements Primitive {
 
     @Override
     public void accept(Visitor visitor) {
